Make network test work
