+++ testing/src/lib.rs
@@ -1,7 +1,7 @@
 use hotshot_task::{event_stream::ChannelStream, task_impls::HSTWithEvent};
 
-/// Initialized system context handle.
-pub mod system_handle;
+/// Helpers for initializing system context handle and building tasks.
+pub mod task_helpers;
 
 ///  builder
 pub mod test_builder;
