+++ testing/tests/consensus_task.rs
@@ -13,6 +13,7 @@ use hotshot_task_impls::events::SequencingHotShotEvent;
 use hotshot_task_impls::harness::run_harness;
 use hotshot_testing::node_types::SequencingMemoryImpl;
 use hotshot_testing::node_types::SequencingTestTypes;
+use hotshot_testing::task_helpers::{build_quorum_proposal, key_pair_for_id};
 use hotshot_types::data::QuorumProposal;
 use hotshot_types::data::SequencingLeaf;
 use hotshot_types::data::ViewNumber;
@@ -27,80 +28,6 @@ use hotshot_types::traits::{
 };
 
 use std::collections::HashMap;
-
-async fn build_proposal(
-    handle: &SystemContextHandle<SequencingTestTypes, SequencingMemoryImpl>,
-    private_key: &<BN254Pub as SignatureKey>::PrivateKey,
-) -> Proposal<QuorumProposal<SequencingTestTypes, SequencingLeaf<SequencingTestTypes>>> {
-    let (proposal, signature) = build_proposal_and_signature(handle, private_key).await;
-    Proposal {
-        data: proposal,
-        signature,
-    }
-}
-
-async fn build_proposal_and_signature(
-    handle: &SystemContextHandle<SequencingTestTypes, SequencingMemoryImpl>,
-    private_key: &<BN254Pub as SignatureKey>::PrivateKey,
-) -> (
-    QuorumProposal<SequencingTestTypes, SequencingLeaf<SequencingTestTypes>>,
-    EncodedSignature,
-) {
-    let consensus_lock = handle.get_consensus();
-    let consensus = consensus_lock.read().await;
-    let api: HotShotSequencingConsensusApi<SequencingTestTypes, SequencingMemoryImpl> =
-        HotShotSequencingConsensusApi {
-            inner: handle.hotshot.inner.clone(),
-        };
-    let _quorum_exchange = api.inner.exchanges.quorum_exchange().clone();
-
-    let parent_view_number = &consensus.high_qc.view_number();
-    let Some(parent_view) = consensus.state_map.get(parent_view_number) else {
-                    panic!("Couldn't find high QC parent in state map.");
-                };
-    let Some(leaf) = parent_view.get_leaf_commitment() else {
-                    panic!(
-                        "Parent of high QC points to a view without a proposal"
-                    );
-                };
-    let Some(leaf) = consensus.saved_leaves.get(&leaf) else {
-                    panic!("Failed to find high QC parent.");
-                };
-    let parent_leaf = leaf.clone();
-
-    // every event input is seen on the event stream in the output.
-
-    let block_commitment = <SequencingTestTypes as NodeType>::BlockType::new().commit();
-    let leaf = SequencingLeaf {
-        view_number: ViewNumber::new(1),
-        height: parent_leaf.height + 1,
-        justify_qc: consensus.high_qc.clone(),
-        parent_commitment: parent_leaf.commit(),
-        // Use the block commitment rather than the block, so that the replica can construct
-        // the same leaf with the commitment.
-        deltas: Right(block_commitment),
-        rejected: vec![],
-        timestamp: 0,
-        proposer_id: api.public_key().to_bytes(),
-    };
-    let signature = <BN254Pub as SignatureKey>::sign(private_key, leaf.commit().as_ref());
-    let proposal = QuorumProposal::<SequencingTestTypes, SequencingLeaf<SequencingTestTypes>> {
-        block_commitment,
-        view_number: ViewNumber::new(1),
-        height: 1,
-        justify_qc: QuorumCertificate::genesis(),
-        proposer_id: leaf.proposer_id,
-        dac: None,
-    };
-
-    (proposal, signature)
-}
-
-fn key_pair_for_id(node_id: u64) -> (<BN254Pub as SignatureKey>::PrivateKey, BN254Pub) {
-    let private_key = <BN254Pub as SignatureKey>::generated_from_seed_indexed([0u8; 32], node_id).1;
-    let public_key = <SequencingTestTypes as NodeType>::SignatureKey::from_private(&private_key);
-    (private_key, public_key)
-}
 
 async fn build_vote(
     handle: &SystemContextHandle<SequencingTestTypes, SequencingMemoryImpl>,
@@ -168,12 +95,12 @@ async fn build_vote(
 #[cfg_attr(feature = "async-std-executor", async_std::test)]
 async fn test_consensus_task() {
     use hotshot_task_impls::harness::build_harness;
-    use hotshot_testing::system_handle::build_system_handle;
+    use hotshot_testing::task_helpers::build_system_handle;
 
     async_compatibility_layer::logging::setup_logging();
     async_compatibility_layer::logging::setup_backtrace();
 
-    let handle = build_system_handle(1).await;
+    let handle = build_system_handle(1).await.0;
     let (private_key, public_key) = key_pair_for_id(1);
 
     let mut input = Vec::new();
@@ -185,7 +112,7 @@ async fn test_consensus_task() {
 
     output.insert(
         SequencingHotShotEvent::QuorumProposalSend(
-            build_proposal(&handle, &private_key).await,
+            build_quorum_proposal(&handle, &private_key,1).await,
             public_key,
         ),
         1,
@@ -194,7 +121,7 @@ async fn test_consensus_task() {
     output.insert(SequencingHotShotEvent::ViewChange(ViewNumber::new(2)), 2);
     output.insert(SequencingHotShotEvent::Shutdown, 1);
 
-    let (task_runner, event_stream) = build_harness(output).await;
+    let (task_runner, event_stream) = build_harness(output, None).await;
     let task_runner = add_consensus_task(
         task_runner,
         event_stream.clone(),
@@ -213,18 +140,18 @@ async fn test_consensus_task() {
 #[cfg_attr(feature = "async-std-executor", async_std::test)]
 async fn test_consensus_vote() {
     use hotshot_task_impls::harness::build_harness;
-    use hotshot_testing::system_handle::build_system_handle;
+    use hotshot_testing::task_helpers::build_system_handle;
 
     async_compatibility_layer::logging::setup_logging();
     async_compatibility_layer::logging::setup_backtrace();
 
-    let handle = build_system_handle(2).await;
+    let handle = build_system_handle(2).await.0;
     let (private_key, public_key) = key_pair_for_id(1);
 
     let mut input = Vec::new();
     let mut output = HashMap::new();
 
-    let proposal = build_proposal(&handle, &private_key).await;
+    let proposal = build_quorum_proposal(&handle, &private_key,1).await;
 
     input.push(SequencingHotShotEvent::ViewChange(ViewNumber::new(1)));
     input.push(SequencingHotShotEvent::QuorumProposalRecv(
@@ -248,7 +175,7 @@ async fn test_consensus_vote() {
     output.insert(SequencingHotShotEvent::ViewChange(ViewNumber::new(2)), 1);
     output.insert(SequencingHotShotEvent::Shutdown, 1);
 
-    let (task_runner, event_stream) = build_harness(output).await;
+    let (task_runner, event_stream) = build_harness(output, None).await;
     let task_runner = add_consensus_task(
         task_runner,
         event_stream.clone(),
